+++ CHANGELOG.md
@@ -72,6 +72,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PendingWindow` is a new type that can return a `WindowHandle` for a window
   that hasn't opened yet. This can be used to allow a widget on a window to
   close the window.
+- Style components for customizing default widget colors have been added:
+  - `DefaultForegroundColor`
+  - `DefaultBackgroundColor`
+  - `DefaultHoveredForegroundColor`
+  - `DefaultHoveredBackgroundColor`
+  - `DefaultActiveForegroundColor`
+  - `DefaultActiveBackgroundColor`
+  - `DefaultDisabledForegroundColor`
+  - `DefaultDisabledBackgroundColor`
 
 [91]: https://github.com/khonsulabs/gooey/issues/91
 [92]: https://github.com/khonsulabs/gooey/issues/92